// Airbnb-front/src/app/components/models/interestedInOnlineExperience.ts
export class interestedInOnlineExperience{
    interestedInPersonName: string;
    interestedInPersonEmail: string;
    onlineExperienceHostEmail: string;
    onlineExperienceNumber: number;
    onlineExperienceType: any;
    onlineExperienceDate: Date;
    onlineExperienceStartTime: Date;
    onlineExperienceEndTime: Date;
    onlineExperienceLink:string
}